♻️ refactor: 프로필 코드 분리
