+++ frontend/src/app/providers/AuthProvider.tsx
@@ -1,6 +1,14 @@
-import { useState, type ReactNode, useCallback, useMemo } from "react";
-import { AuthContext, type UserAccess, type AuthContextValue } from "@/app/providers/auth-context";
-import { getAccessToken, setTokens, clearTokens } from "@/shared/lib/auth-storage";
+import { useState, type ReactNode, useCallback, useMemo } from 'react';
+import {
+  AuthContext,
+  type UserAccess,
+  type AuthContextValue,
+} from '@/app/providers/auth-context';
+import {
+  getAccessToken,
+  setTokens,
+  clearTokens,
+} from '@/shared/lib/auth-storage';
 
 const getUserIdFromToken = (token: string | null): string | null => {
   if (!token) return null;
@@ -8,13 +16,14 @@ const getUserIdFromToken = (token: string | null): string | null => {
     const base64Url = token.split('.')[1];
     const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
     const jsonPayload = decodeURIComponent(
-      window.atob(base64)
+      window
+        .atob(base64)
         .split('')
-        .map(c => '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2))
-        .join('')
+        .map((c) => '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2))
+        .join(''),
     );
     return JSON.parse(jsonPayload).id || null;
-  } catch (e) {
+  } catch {
     return null;
   }
 };
